record block position as 24-bit scid tx index for watched txs (NL-102)
Count every transaction in the block (coinbase = 0) instead of only watched ones, widen TransactionIndex from ushort to uint end to end, and add the WidenWatchedTransactionIndex migration for all three providers.

(cherry picked from commit 519928739304acb1f73e31c71be8be5a6d05f653)

+++ src/NLightning.Infrastructure.Persistence.Sqlite/Migrations/NLightningDbContextModelSnapshot.cs
@@ -121,7 +121,7 @@ namespace NLightning.Infrastructure.Persistence.Sqlite.Migrations
                     b.Property<uint>("RequiredDepth")
                         .HasColumnType("INTEGER");
 
-                    b.Property<ushort?>("TransactionIndex")
+                    b.Property<uint?>("TransactionIndex")
                         .HasColumnType("INTEGER");
 
                     b.HasKey("TransactionId");
